New ChannelController based on Flow
